backend/requirements.txt
aiosmtpd==1.4.5
asgiref==3.8.1
atpublic==4.1.0
attrs==23.2.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
Faker==24.14.0
fastjsonschema==2.19.1
filelock==3.13.4
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
jwcrypto==1.5.6
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pygraphviz==1.12
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rapidfuzz==3.8.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.4
trove-classifiers==2024.4.10
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.25.1
xattr==1.1.0
zipp==3.18.1
django_redis